Remove warning 4 (#3)
For a type like:

```ocaml
type t =
| A
| B
| C of int
```

The deriver would generate a single `_` case in the pattern patching for
`bin_size`. This triggers an instance of warning 4, "fragile pattern
matching". While innocuous in that case, it requires global suppression
or similar tricks.

Instead, this uses a or-case so that the pattern matching is done
extensively.